Rename 'dokument' to 'document' and refactor related code
'MyDokument' is renamed to 'MyDocument', and instances where it was referenced were updated to reflect this change as well. Aside from renaming, some changes in the handling of data were introduced. For instance, in documentParser.py, the file content is now read and decoded in utf-8. Also, the '__init__' file from migrations has been repurposed to 'jaccary.json' which includes the prompt and context data. Minor corrections on conditional statements in views.py is also part of this commit.