Refactor settings and update options.json structure
The commit involved two main changes. Firstly, I introduced the usage of "VECTOR_DB" from the settings in multiple files instead of hard-coding the database path. This promotes better consistency and ease of configuration AND prevent miss configuration as it was before! The README is also update.

Secondly, I refactored the structure of options.json and updated the related tests, which now reflect the new structure, optimizing clarity and organization.