Remove unused test endpoint and update function calls
Removed the redundant "test" endpoint from options.json file as it was merely a placeholder. Also, modified the name of the "rag_function_call" from "jaccard" to "jaccard_index" in the test_views.py file. In addition, the unnecessary "database" attribute was removed from the serializers.py file and an update was made to the response message in views.py within the document directory.